automatically backed up by learn

+++ index.js
@@ -15,7 +15,8 @@ else {
 function currentLine(x) {
   if (x.length > 0) {
    var i = [0, ...x.length]
-  return `The line is currently: ${i + 1}. ${x[i]} `
+  return `The line is currently: ${i.toInt + 1}. ${x[i]} `
+    
   }
   else{
     return "The line is currently empty."
